requirements-dev.txt
-r requirements.txt
pytest
recommonmark
twine
numpydoc
sphinx
sphinx-copybutton
sphinx_rtd_theme
sphinxcontrib-bibtex
